+++ lib/src/annotations/sql_schema.dart
@@ -9,5 +9,12 @@ class SqlSchema {
   /// Can be either a list of strings (column names) or SqlColumn objects
   final List<SqlColumn> columns;
 
-  const SqlSchema({required this.version, required this.columns});
+  /// Optional list of seed data maps
+  final List<Map<String, dynamic>> seedData;
+
+  const SqlSchema({
+    required this.version,
+    required this.columns,
+    this.seedData = const <Map<String, dynamic>>[],
+  });
 }