+++ test/models/new_user.g.dart
@@ -30,6 +30,9 @@ class NewUserTable extends SqlEngineTable {
   };
 
   @override
+  List<Map<String, dynamic>> get initialSeedData => <Map<String, dynamic>>[];
+
+  @override
   List<String> get createIndexes => [];
 
   @override