+++ test/models/order_item.g.dart
@@ -26,6 +26,9 @@ class OrderItemTable extends SqlEngineTable {
   };
 
   @override
+  List<Map<String, dynamic>> get initialSeedData => <Map<String, dynamic>>[];
+
+  @override
   List<String> get createIndexes => [];
 
   @override