+++ test/models/user_mg.g.dart
@@ -29,6 +29,9 @@ class UserMgTable extends SqlEngineTable {
   };
 
   @override
+  List<Map<String, dynamic>> get initialSeedData => <Map<String, dynamic>>[];
+
+  @override
   List<String> get createIndexes => [];
 
   @override